hosts/pi4: remove disko-config
I kept running into boot failures when installing the hosts/pi4 config
via nixos-anywhere.

The reason it kept failing was because disko is unable to change the
partitions and file systems since we're booting from the SD card, and
that is in use by the current system. This results in errors like:

```
Error: Partition(s) 2, 3 on /dev/mmcblk0 have been written, but we have
been unable to inform the kernel of the change, probably because it/they
are in use.  As a result, the old partition(s) will remain in use.
You should reboot now before making further changes.
```

and:

```
mke2fs 1.47.2 (1-Jan-2025)

The file /dev/disk/by-partlabel/disk-main-root does not exist and no
size was specified.
```

I could boot from a USB instead of SD card, but I don't see the reason
to do that because I'm not doing anything fancy with the storage, i.e. I
don't plan on adding an nvme or external storage. So I don't need to
re-partition the drive.

Booting off of the SD card imaged with 'sd-image' works fine, so we can
just use nixos-rebuild instead.

See https://github.com/nvmd/nixos-raspberrypi/issues/8#issuecomment-2804912881