all: remove colmena
Remove colmena since it is too restrictive when it comes to using
different nixpkgs forks, so doesn't work with nixos-raspberrypi.

It is much simpler for me to just use 'nixos-rebuild'. I don't need to
build and deploy on multiple hosts in parallel, so this was an oversight
and over-engineered solution for my requirements from the very
beginning (before I realized I could just use 'nixos-rebuild' from my
MBP).