hosts/pi4: workaround boot-firmware error
When testing the nixos build on the pi4, it will error out with:

reloading user units for nixos...
restarting sysinit-reactivation.target
Error: Failed to open unit file /nix/store/.../etc/systemd/system/boot-firmware.mount

This happens because the system is looking for boot-firmware.mount unit
that doesn't exist. It looks due to a timing / activation issue with
systemd automounts, and since my pi4 is using a 32GB SD card and only
has 4GB of RAM to work with, it might be too slow.

Instead, it's easier to just use the default generated hardware config.

Although I'm not sure this only worked after:
> sudo mount /dev/disk/by-label/FIRMWARE /boot/firmware

But I left the note in the build script for posterity.