scripts/smoketests: add gateway smoketests
Add test-caddy to test that caddy redirects to the expected hosts and
we're able to successfully connect to the redirected host.

There's a discrepancy between how dig/nslookup and curl/ping resolve
domains on OSX, so using the '--resolve' flag to bypass the system
resolver. That's something I still need to look into to fix :').