experimental: Use docker-compose for multi-arch builds
The goal of this experiment was to use local docker builders for
building multi-arch packages / images.

In the end, I was able to get the headless builders up and running and
connect over ssh, but I was unable to get nix build to connect to them.

I tried NIX_SSHOPTS, builders file, and (failed to) override the `ssh`
binary with a wrapper script to get some debug info. I continued to run
into `Host key verification failure` and seems like NIX_SSHOPTS were
being straight-up ignored, which didn't follow the actual nix source
code.

Anyway, this was an interesting side quest, but I'm giving up on it.